speculation-rules/prefetch/prefetch-status-updated.html

<!DOCTYPE html>
<meta name="timeout" content="long">
<script src="/resources/testharness.js"></script>
<script src="/resources/testharnessreport.js"></script>
<script src="/resources/testdriver.js?feature=bidi"></script>
<script src="/common/utils.js"></script>
<script src="/common/dispatcher/dispatcher.js"></script>
<script src="../resources/utils.js"></script>
<script src="resources/utils.sub.js"></script>

<meta name="variant" content="?same-site">
<meta name="variant" content="?cross-site">

<script>
setup(() => assertSpeculationRulesIsSupported());

promise_test(async t => {
  // Subscribe to prefetch status updates
  await test_driver.bidi.speculation.prefetch_status_updated.subscribe();
  
  // Array to collect all status events
  const statusEvents = [];
  
  // Set up event listener to capture all prefetch status updates
  const removeListener = test_driver.bidi.speculation.prefetch_status_updated.on((event) => {
    console.log('Prefetch status event received:', event);
    statusEvents.push(event);
  });
  
  try {
    // Create a test window
    const agent = await spawnWindow(t);
    const nextUrl = agent.getExecutorURL({ page: 2 });
    
    console.log(`Starting prefetch for: ${nextUrl}`);
    
    // Start the prefetch - this just adds speculation rules
    await agent.forceSinglePrefetch(nextUrl);
    
    // Now wait for prefetch status events to be received
    // We'll wait up to 3 seconds for events to arrive
    const maxWaitTime = 3000;
    const startTime = Date.now();
    
    while (statusEvents.length === 0 && (Date.now() - startTime) < maxWaitTime) {
      await new Promise(resolve => setTimeout(resolve, 100));
    }
    
    console.log(`Waited ${Date.now() - startTime}ms for events`);
    
    // Navigate to the prefetched URL
    await agent.navigate(nextUrl);
    
    // Log the results
    console.log(`Total prefetch status events received: ${statusEvents.length}`);
    
    if (statusEvents.length > 0) {
      statusEvents.forEach((event, index) => {
        console.log(`Event ${index + 1}:`, {
          url: event.url || 'unknown',
          status: event.status || 'unknown',
          initiator: event.initiator || 'unknown',
          timestamp: event.timestamp || 'unknown',
          error: event.error || 'none'
        });
      });
      
      // Check if we got a successful prefetch status
      const hasSuccessStatus = statusEvents.some(event => 
        event.status === 'success' || event.status === 'complete'
      );
      
      if (hasSuccessStatus) {
        console.log('✓ Prefetch completed successfully');
      } else {
        console.log('⚠ No successful prefetch status detected');
      }
      
      // Basic assertion - we should have received at least one status event
      assert_greater_than(statusEvents.length, 0, 'Should receive at least one prefetch status event');
    } else {
      console.log('ℹ No prefetch status events received within timeout');
      console.log('This may be expected if the browser implementation doesn\'t emit these events yet');
      
      assert_unreached('Expected to receive prefetch status events');
    }
    
  } finally {
    // Clean up the event listener
    removeListener();
  }
}, 'Test that prefetch_status_updated events are properly received and logged.');
</script>
